refactor(widget-plugin-test-utils): move functions file
